Only build the Ruby plugin on Fedora >= 30
Fedora < 30 have protobuf 3.5, which doesn't have the required parts to
build the Ruby plugin.

We can't update protobuf there, because upstream bumped the soname which
goes contrary to the Fedora update policy.

As a result, if we want Fedora 28 and 29 to have gRPC, we have to
disable its Ruby plugin.

Upstream doesn't seem to have an easy way to just not build it, so this
commit introduces a patch to remove it from the Makefile.

This should be okay, since it's only temporary anyway: Fedora 30 (and
future releases) already have a recent enough protobuf so this won't be
necessary any more.