Use C++11 during compilation
As we built abseil-cpp with C++11, we need to set it
11 here as well otherwise we have linker errors.

We use pkgconfig() instead of cmake() for c-ares as
the latter one doesn't work.

We requires protobuf 3.12 as per [1].

Plus, we need to use the directories of CA files as
argument for symlinks. The version we have currently
in CS8-BaseOS (i.e 1.4) expects only directories.

[1] https://github.com/grpc/grpc/pull/23324

Change-Id: I7c3a946c1c4099b14dac25003aa594e2c014d4c9